RN-410 display two different types of overlays separately  (#3803)
